Update OrderWords.js
- Fixed a bug where isolated "-" tokens would appear in the word-soup. This is mostly due to articles where there is a leading dash.
- Fixed a bug where parenthesis would cause the reduce context to fail as they are read as regex characters rather than string. To avoid this, the string is escaped before using .search.
- The properties "sentence_context_was_reduced" and "sentence_was_too_long" were a bit confusing. So they were changed to better encapsulate their meaning.
"sentence_was_too_long" -> This is true when the context reducer was used, so it is named "was_context_reduce_used"
"sentence_context_was_reduced" -> True when the exercise is handling long sentences. This is expected to be true at all times, but it can be useful when disabling the context reducer for other tests.